ファイル読み取り処理、ファイル削除時にファイルの存在を確認

+++ api/config/localfile.js
@@ -1,6 +1,13 @@
 const fs = require('fs');
 
 module.exports = {
+  // ファイル読み取り処理
+  getFiles(filePath) {
+    fs.readFile(filePath, (err, data) => {
+      if (err) throw err;
+      return data;
+    });
+  },
   // ファイル書き込み処理
   uploadFiles(fileList, username) {
     fileList.map(file => {
@@ -9,7 +16,7 @@ module.exports = {
       if (!fs.existsSync(path)) {
         fs.mkdirSync(path);
       }
-    
+      
       fs.renameSync(
         file.path,
         path + file.filename + "." +
@@ -19,6 +26,8 @@ module.exports = {
   },
   // ファイル削除処理
   deleteFiles(filePath) {
-    fs.unlinkSync(filePath);
+    if (ts.existsSync(filePath)) {
+      fs.unlinkSync(filePath);
+    }
   }
-}
+};
